fix datasource blockchain_platforms

+++ oci/blockchain_osn_resource.go
@@ -330,7 +330,7 @@ func (s *BlockchainOsnResourceCrud) mapToOcpuAllocationNumberParam(fieldKeyForma
 	result := oci_blockchain.OcpuAllocationNumberParam{}
 
 	if ocpuAllocationNumber, ok := s.D.GetOkExists(fmt.Sprintf(fieldKeyFormat, "ocpu_allocation_number")); ok {
-		tmp := ocpuAllocationNumber.(float32)
+		tmp := float32(ocpuAllocationNumber.(float64))
 		result.OcpuAllocationNumber = &tmp
 	}
